Add displayallcomments functoin & styles for comment-ul
