+++ src/style.css
@@ -86,7 +86,7 @@ main {
 }
 
 .comment-section h2 {
-  color: white;
+  color: black;
   text-align: center;
   font-size: 20px;
   padding-bottom: 10px;
@@ -98,6 +98,26 @@ main {
   align-items: center;
 }
 
+.user,
+.comment {
+  padding: 5px;
+  border-radius: 8px;
+}
+
+.comment-ul {
+  flex-direction: column;
+  list-style: none;
+}
+
+.comment-ul li {
+  background-color: wheat;
+  padding: 12px;
+  border-radius: 8px;
+  margin-top: 20px;
+  width: auto;
+  color: black;
+}
+
 .submitBtn {
   display: flex;
   justify-content: center;
@@ -107,12 +127,6 @@ main {
   text-align: center;
   cursor: pointer;
   box-shadow: 3px 3px 8px 1px rgb(0 0 0 / 35%);
-}
-
-.user,
-.comment {
-  padding: 5px;
-  border-radius: 8px;
 }
 
 @media only screen and (max-width: 600px) {
